Add files via upload
The "toString()" tests were failing because I did not type the actual message correctly. These files are the files without the typos. These typos were causing me to lose points in the ZyLabs grading system. 

Also, I misread the lab and put in tests for every single one of the getter methods in the Zoo class. So I removed those lines of code that tested "getAverageWeight(String color)" or "getAnimals" for the sake of following the project's examples.